Color Changes, Portal, and Spikes
I created code for the obj_portal, obj_spikes, and color changes. I am
attempting to make obj_player hold most to all of the code so that's the
only object we really have to worry about if there is an error and not
have to jump between objects to attempt and find the problem.